+++ src/lib.rs
@@ -6,7 +6,7 @@
 //! The library is aimed at ease-of-use and performance first.
 //!
 //! ```
-//! use fugit::{Duration, ExtU32};
+//! use fugit::{Duration, DurationExtU32};
 //!
 //! // Efficient short-hands (`.millis()`, ...)
 //! let d = Duration::<u32, 1, 1_000>::from_ticks(111);
@@ -48,15 +48,24 @@ mod aliases;
 mod duration;
 mod helpers;
 mod instant;
+mod rate;
 
 pub use aliases::*;
-pub use duration::{Duration, ExtU32, ExtU64};
+pub use duration::{Duration, ExtU32 as DurationExtU32, ExtU64 as DurationExtU64};
 pub use instant::Instant;
+pub use rate::{ExtU32 as RateExtU32, ExtU64 as RateExtU64, Rate};
 
 #[cfg(test)]
 mod test {
     use crate::Duration;
     use crate::Instant;
+    use crate::Rate;
+
+    ////////////////////////////////////////////////////////////////////////////////
+    //
+    // Duration tests
+    //
+    ////////////////////////////////////////////////////////////////////////////////
 
     fn take_ms(d: Duration<u32, 1, 1_000>) -> Duration<u32, 1, 1_000> {
         d
@@ -428,7 +437,7 @@ mod test {
 
     #[test]
     fn duration_duration_math_u32() {
-        use crate::ExtU32;
+        use crate::DurationExtU32;
 
         // Same base
         let sum: Duration<u32, 1, 1_000> =
@@ -459,7 +468,7 @@ mod test {
 
     #[test]
     fn duration_duration_math_u64() {
-        use crate::ExtU64;
+        use crate::DurationExtU64;
 
         // Same base
         let sum: Duration<u64, 1, 1_000> =
@@ -507,7 +516,7 @@ mod test {
 
     #[test]
     fn duration_shorthands_u32() {
-        use crate::ExtU32;
+        use crate::DurationExtU32;
 
         let d: Duration<u32, 1, 10_000> = 100_000.micros();
         assert_eq!(d.ticks(), 1_000);
@@ -527,7 +536,7 @@ mod test {
 
     #[test]
     fn duration_shorthands_u64() {
-        use crate::ExtU64;
+        use crate::DurationExtU64;
 
         let d: Duration<u64, 1, 10_000> = 100_000.micros();
         assert_eq!(d.ticks(), 1_000);
@@ -544,6 +553,12 @@ mod test {
         let d: Duration<u64, 1, 10_000> = 1.hours();
         assert_eq!(d.ticks(), 36_000_000);
     }
+
+    ////////////////////////////////////////////////////////////////////////////////
+    //
+    // Instant tests
+    //
+    ////////////////////////////////////////////////////////////////////////////////
 
     #[test]
     fn instant_compare_u32() {
@@ -651,7 +666,7 @@ mod test {
 
     #[test]
     fn instant_duration_math_u32() {
-        use crate::ExtU32;
+        use crate::DurationExtU32;
 
         // Instant - Instant, Same base
         let diff: Duration<u32, 1, 1_000> =
@@ -687,7 +702,7 @@ mod test {
 
     #[test]
     fn instant_duration_math_u64() {
-        use crate::ExtU64;
+        use crate::DurationExtU64;
 
         // Instant - Instant, Same base
         let diff: Duration<u64, 1, 1_000> =
@@ -720,4 +735,232 @@ mod test {
         let diff: Instant<u64, 1, 10_000> = Instant::<u64, 1, 10_000>::from_ticks(10) - 1.millis();
         assert_eq!(diff, Instant::<u64, 1, 10_000>::from_ticks(0));
     }
+
+    ////////////////////////////////////////////////////////////////////////////////
+    //
+    // Rate tests
+    //
+    ////////////////////////////////////////////////////////////////////////////////
+
+    fn take_khz(r: Rate<u32, 1_000, 1>) -> Rate<u32, 1_000, 1> {
+        r
+    }
+
+    #[test]
+    fn rate_functions() {
+        assert_eq!(
+            take_khz(Rate::<u32, 10_000, 1>::from_ticks(1).convert()),
+            Rate::<u32, 1_000, 1>::from_ticks(10)
+        );
+    }
+
+    #[test]
+    fn rate_compare_u32() {
+        // Same fraction
+        assert!(Rate::<u32, 1_000, 1>::from_ticks(2) > Rate::<u32, 1_000, 1>::from_ticks(1));
+        assert!(Rate::<u32, 1_000, 1>::from_ticks(2) >= Rate::<u32, 1_000, 1>::from_ticks(1));
+        assert!(Rate::<u32, 1_000, 1>::from_ticks(1) >= Rate::<u32, 1_000, 1>::from_ticks(1));
+        assert!(Rate::<u32, 1_000, 1>::from_ticks(1) < Rate::<u32, 1_000, 1>::from_ticks(2));
+        assert!(Rate::<u32, 1_000, 1>::from_ticks(1) <= Rate::<u32, 1_000, 1>::from_ticks(1));
+        assert!(Rate::<u32, 1_000, 1>::from_ticks(1) <= Rate::<u32, 1_000, 1>::from_ticks(2));
+        assert!(Rate::<u32, 1_000, 1>::from_ticks(1) == Rate::<u32, 1_000, 1>::from_ticks(1));
+        assert!(Rate::<u32, 1_000, 1>::from_ticks(1) != Rate::<u32, 1_000, 1>::from_ticks(2));
+
+        // Different fraction
+        assert!(Rate::<u32, 1_000, 1>::from_ticks(11) > Rate::<u32, 10_000, 1>::from_ticks(1));
+        assert!(Rate::<u32, 1_000, 1>::from_ticks(11) >= Rate::<u32, 10_000, 1>::from_ticks(1));
+        assert!(Rate::<u32, 1_000, 1>::from_ticks(10) >= Rate::<u32, 10_000, 1>::from_ticks(1));
+        assert!(Rate::<u32, 1_000, 1>::from_ticks(11) < Rate::<u32, 10_000, 1>::from_ticks(2));
+        assert!(Rate::<u32, 1_000, 1>::from_ticks(1) <= Rate::<u32, 10_000, 1>::from_ticks(1));
+        assert!(Rate::<u32, 1_000, 1>::from_ticks(10) <= Rate::<u32, 10_000, 1>::from_ticks(1));
+        assert!(Rate::<u32, 1_000, 1>::from_ticks(10) == Rate::<u32, 10_000, 1>::from_ticks(1));
+        assert!(Rate::<u32, 1_000, 1>::from_ticks(9) != Rate::<u32, 10_000, 1>::from_ticks(2));
+    }
+
+    #[test]
+    fn rate_compare_u64() {
+        // Same fraction
+        assert!(Rate::<u64, 1_000, 1>::from_ticks(2) > Rate::<u64, 1_000, 1>::from_ticks(1));
+        assert!(Rate::<u64, 1_000, 1>::from_ticks(2) >= Rate::<u64, 1_000, 1>::from_ticks(1));
+        assert!(Rate::<u64, 1_000, 1>::from_ticks(1) >= Rate::<u64, 1_000, 1>::from_ticks(1));
+        assert!(Rate::<u64, 1_000, 1>::from_ticks(1) < Rate::<u64, 1_000, 1>::from_ticks(2));
+        assert!(Rate::<u64, 1_000, 1>::from_ticks(1) <= Rate::<u64, 1_000, 1>::from_ticks(1));
+        assert!(Rate::<u64, 1_000, 1>::from_ticks(1) <= Rate::<u64, 1_000, 1>::from_ticks(2));
+        assert!(Rate::<u64, 1_000, 1>::from_ticks(1) == Rate::<u64, 1_000, 1>::from_ticks(1));
+        assert!(Rate::<u64, 1_000, 1>::from_ticks(1) != Rate::<u64, 1_000, 1>::from_ticks(2));
+
+        // Different fraction
+        assert!(Rate::<u64, 1_000, 1>::from_ticks(11) > Rate::<u64, 10_000, 1>::from_ticks(1));
+        assert!(Rate::<u64, 1_000, 1>::from_ticks(11) >= Rate::<u64, 10_000, 1>::from_ticks(1));
+        assert!(Rate::<u64, 1_000, 1>::from_ticks(10) >= Rate::<u64, 10_000, 1>::from_ticks(1));
+        assert!(Rate::<u64, 1_000, 1>::from_ticks(11) < Rate::<u64, 10_000, 1>::from_ticks(2));
+        assert!(Rate::<u64, 1_000, 1>::from_ticks(1) <= Rate::<u64, 10_000, 1>::from_ticks(1));
+        assert!(Rate::<u64, 1_000, 1>::from_ticks(10) <= Rate::<u64, 10_000, 1>::from_ticks(1));
+        assert!(Rate::<u64, 1_000, 1>::from_ticks(10) == Rate::<u64, 10_000, 1>::from_ticks(1));
+        assert!(Rate::<u64, 1_000, 1>::from_ticks(9) != Rate::<u64, 10_000, 1>::from_ticks(2));
+    }
+
+    #[test]
+    fn rate_compare_u64_u32() {
+        // Same fraction
+        assert!(Rate::<u64, 1_000, 1>::from_ticks(2) > Rate::<u32, 1_000, 1>::from_ticks(1));
+        assert!(Rate::<u64, 1_000, 1>::from_ticks(2) >= Rate::<u32, 1_000, 1>::from_ticks(1));
+        assert!(Rate::<u64, 1_000, 1>::from_ticks(1) >= Rate::<u32, 1_000, 1>::from_ticks(1));
+        assert!(Rate::<u64, 1_000, 1>::from_ticks(1) < Rate::<u32, 1_000, 1>::from_ticks(2));
+        assert!(Rate::<u64, 1_000, 1>::from_ticks(1) <= Rate::<u32, 1_000, 1>::from_ticks(1));
+        assert!(Rate::<u64, 1_000, 1>::from_ticks(1) <= Rate::<u32, 1_000, 1>::from_ticks(2));
+        assert!(Rate::<u64, 1_000, 1>::from_ticks(1) == Rate::<u32, 1_000, 1>::from_ticks(1));
+        assert!(Rate::<u64, 1_000, 1>::from_ticks(1) != Rate::<u32, 1_000, 1>::from_ticks(2));
+
+        // Different fraction
+        assert!(Rate::<u64, 1_000, 1>::from_ticks(11) > Rate::<u32, 10_000, 1>::from_ticks(1));
+        assert!(Rate::<u64, 1_000, 1>::from_ticks(11) >= Rate::<u32, 10_000, 1>::from_ticks(1));
+        assert!(Rate::<u64, 1_000, 1>::from_ticks(10) >= Rate::<u32, 10_000, 1>::from_ticks(1));
+        assert!(Rate::<u64, 1_000, 1>::from_ticks(11) < Rate::<u32, 10_000, 1>::from_ticks(2));
+        assert!(Rate::<u64, 1_000, 1>::from_ticks(1) <= Rate::<u32, 10_000, 1>::from_ticks(1));
+        assert!(Rate::<u64, 1_000, 1>::from_ticks(10) <= Rate::<u32, 10_000, 1>::from_ticks(1));
+        assert!(Rate::<u64, 1_000, 1>::from_ticks(10) == Rate::<u32, 10_000, 1>::from_ticks(1));
+        assert!(Rate::<u64, 1_000, 1>::from_ticks(9) != Rate::<u32, 10_000, 1>::from_ticks(2));
+    }
+
+    #[test]
+    fn rate_compare_u32_u64() {
+        // Same fraction
+        assert!(Rate::<u32, 1_000, 1>::from_ticks(2) > Rate::<u64, 1_000, 1>::from_ticks(1));
+        assert!(Rate::<u32, 1_000, 1>::from_ticks(2) >= Rate::<u64, 1_000, 1>::from_ticks(1));
+        assert!(Rate::<u32, 1_000, 1>::from_ticks(1) >= Rate::<u64, 1_000, 1>::from_ticks(1));
+        assert!(Rate::<u32, 1_000, 1>::from_ticks(1) < Rate::<u64, 1_000, 1>::from_ticks(2));
+        assert!(Rate::<u32, 1_000, 1>::from_ticks(1) <= Rate::<u64, 1_000, 1>::from_ticks(1));
+        assert!(Rate::<u32, 1_000, 1>::from_ticks(1) <= Rate::<u64, 1_000, 1>::from_ticks(2));
+        assert!(Rate::<u32, 1_000, 1>::from_ticks(1) == Rate::<u64, 1_000, 1>::from_ticks(1));
+        assert!(Rate::<u32, 1_000, 1>::from_ticks(1) != Rate::<u64, 1_000, 1>::from_ticks(2));
+
+        // Different fraction
+        assert!(Rate::<u32, 1_000, 1>::from_ticks(11) > Rate::<u64, 10_000, 1>::from_ticks(1));
+        assert!(Rate::<u32, 1_000, 1>::from_ticks(11) >= Rate::<u64, 10_000, 1>::from_ticks(1));
+        assert!(Rate::<u32, 1_000, 1>::from_ticks(10) >= Rate::<u64, 10_000, 1>::from_ticks(1));
+        assert!(Rate::<u32, 1_000, 1>::from_ticks(11) < Rate::<u64, 10_000, 1>::from_ticks(2));
+        assert!(Rate::<u32, 1_000, 1>::from_ticks(1) <= Rate::<u64, 10_000, 1>::from_ticks(1));
+        assert!(Rate::<u32, 1_000, 1>::from_ticks(10) <= Rate::<u64, 10_000, 1>::from_ticks(1));
+        assert!(Rate::<u32, 1_000, 1>::from_ticks(10) == Rate::<u64, 10_000, 1>::from_ticks(1));
+        assert!(Rate::<u32, 1_000, 1>::from_ticks(9) != Rate::<u64, 10_000, 1>::from_ticks(2));
+    }
+
+    #[test]
+    fn rate_rate_math_u32() {
+        use crate::RateExtU32;
+
+        // Same base
+        let sum: Rate<u32, 1_000, 1> =
+            Rate::<u32, 1_000, 1>::from_ticks(10) + Rate::<u32, 1_000, 1>::from_ticks(1);
+        assert_eq!(sum, Rate::<u32, 1_000, 1>::from_ticks(11));
+
+        let diff: Rate<u32, 1_000, 1> =
+            Rate::<u32, 1_000, 1>::from_ticks(10) - Rate::<u32, 1_000, 1>::from_ticks(1);
+        assert_eq!(diff, Rate::<u32, 1_000, 1>::from_ticks(9));
+
+        // Different base
+        let sum: Rate<u32, 10_000, 1> = Rate::<u32, 10_000, 1>::from_ticks(10)
+            + Rate::<u32, 1_000, 1>::from_ticks(10).convert();
+        assert_eq!(sum, Rate::<u32, 10_000, 1>::from_ticks(11));
+
+        let diff: Rate<u32, 10_000, 1> = Rate::<u32, 10_000, 1>::from_ticks(10)
+            - Rate::<u32, 1_000, 1>::from_ticks(10).convert();
+        assert_eq!(diff, Rate::<u32, 10_000, 1>::from_ticks(9));
+
+        // Short hand vs u32 (should not need `.into()`)
+        // let sum = Rate::<u32, 1, 10_000>::from_ticks(10) + 1.millis();
+        // assert_eq!(sum, Rate::<u32, 1, 10_000>::from_ticks(20));
+    }
+
+    #[test]
+    fn rate_rate_math_u64() {
+        use crate::RateExtU64;
+
+        // Same base
+        let sum: Rate<u64, 1_000, 1> =
+            Rate::<u64, 1_000, 1>::from_ticks(10) + Rate::<u64, 1_000, 1>::from_ticks(1);
+        assert_eq!(sum, Rate::<u64, 1_000, 1>::from_ticks(11));
+
+        let diff: Rate<u64, 1_000, 1> =
+            Rate::<u64, 1_000, 1>::from_ticks(10) - Rate::<u64, 1_000, 1>::from_ticks(1);
+        assert_eq!(diff, Rate::<u64, 1_000, 1>::from_ticks(9));
+
+        // Different base
+        let sum: Rate<u64, 10_000, 1> = Rate::<u64, 10_000, 1>::from_ticks(10)
+            + Rate::<u64, 1_000, 1>::from_ticks(10).convert();
+        assert_eq!(sum, Rate::<u64, 10_000, 1>::from_ticks(11));
+
+        let diff: Rate<u64, 10_000, 1> = Rate::<u64, 10_000, 1>::from_ticks(10)
+            - Rate::<u64, 1_000, 1>::from_ticks(10).convert();
+        assert_eq!(diff, Rate::<u64, 10_000, 1>::from_ticks(9));
+
+        // Short hand vs u64 (should not need `.into()`)
+        // let sum = Rate::<u64, 1, 10_000>::from_ticks(10) + 1.millis();
+        // assert_eq!(sum, Rate::<u64, 1, 10_000>::from_ticks(20));
+    }
+
+    #[test]
+    fn rate_rate_math_u64_u32() {
+        // Same base
+        let sum: Rate<u64, 1_000, 1> =
+            Rate::<u64, 1_000, 1>::from_ticks(10) + Rate::<u32, 1_000, 1>::from_ticks(1);
+        assert_eq!(sum, Rate::<u64, 1_000, 1>::from_ticks(11));
+
+        let diff: Rate<u64, 1_000, 1> =
+            Rate::<u64, 1_000, 1>::from_ticks(10) - Rate::<u32, 1_000, 1>::from_ticks(1);
+        assert_eq!(diff, Rate::<u64, 1_000, 1>::from_ticks(9));
+
+        // Different base
+        let sum: Rate<u64, 10_000, 1> = Rate::<u64, 10_000, 1>::from_ticks(10)
+            + Rate::<u32, 1_000, 1>::from_ticks(10).convert();
+        assert_eq!(sum, Rate::<u64, 10_000, 1>::from_ticks(11));
+
+        let diff: Rate<u64, 10_000, 1> = Rate::<u64, 10_000, 1>::from_ticks(10)
+            - Rate::<u32, 1_000, 1>::from_ticks(10).convert();
+        assert_eq!(diff, Rate::<u64, 10_000, 1>::from_ticks(9));
+    }
+
+    #[test]
+    fn rate_shorthands_u32() {
+        use crate::RateExtU32;
+
+        // let d: Rate<u32, 1, 10_000> = 100_000.micros();
+        // assert_eq!(d.ticks(), 1_000);
+
+        // let d: Rate<u32, 1, 10_000> = 1.millis();
+        // assert_eq!(d.ticks(), 10);
+
+        // let d: Rate<u32, 1, 10_000> = 1.secs();
+        // assert_eq!(d.ticks(), 10_000);
+
+        // let d: Rate<u32, 1, 10_000> = 1.minutes();
+        // assert_eq!(d.ticks(), 600_000);
+
+        // let d: Rate<u32, 1, 10_000> = 1.hours();
+        // assert_eq!(d.ticks(), 36_000_000);
+    }
+
+    #[test]
+    fn rate_shorthands_u64() {
+        use crate::RateExtU64;
+
+        // let d: Rate<u64, 1, 10_000> = 100_000.micros();
+        // assert_eq!(d.ticks(), 1_000);
+
+        // let d: Rate<u64, 1, 10_000> = 1.millis();
+        // assert_eq!(d.ticks(), 10);
+
+        // let d: Rate<u64, 1, 10_000> = 1.secs();
+        // assert_eq!(d.ticks(), 10_000);
+
+        // let d: Rate<u64, 1, 10_000> = 1.minutes();
+        // assert_eq!(d.ticks(), 600_000);
+
+        // let d: Rate<u64, 1, 10_000> = 1.hours();
+        // assert_eq!(d.ticks(), 36_000_000);
+    }
+
+    #[test]
+    fn rate_duration_conversion() {}
 }